Correção de Bugs e Validação de inputs vazios e email

+++ controller/consultaControler.js
@@ -1,13 +1,38 @@
+function validacaoEmail(email) {
+    usuario = email.value.substring(0, email.value.indexOf("@"));
+    dominio = email.value.substring(email.value.indexOf("@")+ 1, email.value.length);
+    
+    if ((usuario.length >=1) &&
+        (dominio.length >=3) &&
+        (usuario.search("@")==-1) &&
+        (dominio.search("@")==-1) &&
+        (usuario.search(" ")==-1) &&
+        (dominio.search(" ")==-1) &&
+        (dominio.search(".")!=-1) &&
+        (dominio.indexOf(".") >=1)&&
+        (dominio.lastIndexOf(".") < dominio.length - 1)) {
+            let efocus = document.createElement('p');
+            efocus.setAttribute('id', 'efocus');
+            efocus.style.color = 'green';
+            efocus.innerHTML = 'Email valido!';
+            document.getElementById('valido').appendChild(efocus);
+    }
+    else{
+        $('#tituloErro').html('Email Invalido!');
+        $('#msgErro').html(`Preencha o Email corretamente com @ e .com`);
+        $('#showModalError').modal('show');
+    }
+    }
 /*controler*/
 function pegaDados(id){
     let db = new Db();
     let dados = db.consultar(id);
-    console.log(dados)
+
     if(dados !== null){
         $('#rnome').val(dados[0]._nome).prop('disabled', false).focus();
         $('#rsnome').val(dados[0]._snome).prop('disabled', false);
         $('#rdatanasc').val(dados[0]._datanasc).prop('disabled', false);
-        $('#rcpf').val(dados[0]._id);
+        $('#rcpf').val(dados[0]._cpf);
         $('#rcel').val(dados[0]._cel).prop('disabled', false);
         $('#rg').val(dados[0]._rg)
         $('#rsex').val(dados[0]._sex).prop('disabled', false);
@@ -28,8 +53,150 @@ function pegaDados(id){
         $('#showModalError').modal('show');
     }
 }
+
+
 function validaInput(cpf,nome,snome,datanasc,cel,rg,sex,estciv,cep,rua,bairro,city,est,id,senha){
-    let variaveis = [cpf,nome,snome,datanasc,cel,rg,sex,estciv,cep,rua,bairro,city,est,id,senha]
+    if(cpf == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rcpf').focus()
+        })
+        return false
+    }
+    if(nome == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rnome').focus()
+        })
+        return false
+    }
+    if(snome == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rsnome').focus()
+        })
+        return false
+    }
+    if(datanasc == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rdatanasc').focus()
+        })
+        return false
+    }
+    if(cel == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rcel').focus()
+        })
+        return false
+    }
+    if(rg == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rg').focus()
+        })
+        return false
+    }
+    if(sex == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rsex').focus()
+        })
+        return false
+    }
+    if(estciv == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#restciv').focus()
+        })
+        return false
+    }
+    if(cep == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rcep').focus()
+        })
+        return false
+    }
+    if(rua == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rrua').focus()
+        })
+        return false
+    }
+    if(bairro == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rbairro').focus()
+        })
+        return false
+    }
+    if(city == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rcity').focus()
+        })
+        return false
+    }
+    if(est == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#rest').focus()
+        })
+        return false
+    }
+    if(id == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#id').focus()
+        })
+        return false
+    }
+    if(senha == ''){
+        $('#tituloErro').html('Campo Obrigatorio!');
+        $('#msgErro').html('Verifique se o campo foi preenchido corretamente!')
+        $('#showModalError').modal('show');
+        $('#btnErrorModal').on('click',function(){
+            $('#senha').focus()
+        })
+        return false
+    }
+    return true;
+}
+
+/*
+function validaInput(...variaveis){
+    console.log(variaveis)
     for(let i=0 ; i <= variaveis.length ; i++){
         console.log(variaveis[i])
         if(variaveis[i] == '' || variaveis[i] == null){
@@ -37,15 +204,15 @@ function validaInput(cpf,nome,snome,datanasc,cel,rg,sex,estciv,cep,rua,bairro,ci
           $('#msgErro').html('Verifique se todos os campos foram preenchidos!!')
           $('#showModalError').modal('show');
           return false;
+
         }else{
             
             return true;
         }
     }
 }
-/*
-controler
 */
+
 function criarCadastro(cpf,nome,snome,datanasc,cel,rg,sex,estciv,cep,rua,num,comp,bairro,city,est,id,senha){
      
     const cadastro = []
@@ -61,14 +228,16 @@ function criarCadastro(cpf,nome,snome,datanasc,cel,rg,sex,estciv,cep,rua,num,com
     db.gravar(id, cadastro)
     limparCampos()
     $('#tituloSuccess').html('Usuario Cadastrado!');
-    $('#msgSuccess').html(`Seu Usuario é:${id} \n Sua Senha é: \n ${senha}`);
+    $('#msgSuccess').html(`Seu Usuario é: ${id} \n Sua Senha é: \n ${senha}`);
     $('#showModalSuccess').modal('show');
 }
 
 function excluirCadastro(id){
-    let db = new Db()
-    db.delete(id)
-    limparCampos()
+    let db = new Db();
+    if(id !== ''){
+        db.delete(id)
+        limparCampos()
+    }
 }
 
 function buscaCep(cep){
